# internal/http/docs/swagger.yaml
definitions:
  dto.CreatePasswordCardDTO:
    properties:
      name:
        type: string
      password:
        minLength: 8
        type: string
      url:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - url
    - username
    type: object
  dto.UpdatePasswordCardDTO:
    properties:
      name:
        type: string
      password:
        minLength: 8
        type: string
      url:
        type: string
      username:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      message:
        type: string
      statusCode:
        type: integer
    type: object
  response.ListResponseWithoutGenerics:
    properties:
      data:
        items: {}
        type: array
    type: object
info:
  contact: {}
paths:
  /password-cards:
    get:
      consumes:
      - application/json
      description: List all password cards
      parameters:
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListResponseWithoutGenerics'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List password cards
      tags:
      - PasswordCards
    post:
      consumes:
      - application/json
      description: Create a new password card
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePasswordCardDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create password card
      tags:
      - PasswordCards
  /password-cards/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a password card
      parameters:
      - description: Password Card identification
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete password card
      tags:
      - PasswordCards
    put:
      consumes:
      - application/json
      description: Update a password card
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordCardDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update password card
      tags:
      - PasswordCards
swagger: "2.0"